+++ frontend/src/components/users/userAccount/UserAccount.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { Card, Typography } from "@mui/material";
-import { queryMe } from "@/graphql/Users";
+import { queryMe } from "@/graphql/users/queryMe";
 import { useQuery } from "@apollo/client";
 import { UserTypes } from "@/types/UserTypes";
