--- frontend/src/graphql/Users.ts
@@ -1,83 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const mutationCreateUser = gql`
-  mutation userCreate($data: UserCreateInput!) {
-    item: userCreate(data: $data) {
-      id
-      nickName
-    }
-  }
-`;
-
-export const mutationVerifyEmail = gql`
-  mutation verifyEmail($token: String!) {
-    item: verifyEmail(token: $token) {
-      success
-      message
-    }
-  }
-`;
-
-export const mutationUserLogin = gql`
-  mutation userLogin($data: UserLoginInput!) {
-    item: userLogin(data: $data) {
-      id
-      firstName
-    }
-  }
-`;
-
-export const mutationResetPassword = gql`
-  mutation resetPassword($email: String!) {
-    resetPassword(email: $email)
-  }
-`;
-
-export const mutationSetPassword = gql`
-  mutation setPassword($password: String!, $token: String!) {
-    setPassword(password: $password, token: $token)
-  }
-`;
-
-export const queryMe = gql`
-  query me {
-    item: me {
-      id
-      firstName
-      lastName
-      nickName
-      email
-      picture {
-        id
-        filename
-      }
-      adress
-      zipCode
-      city
-      coordinates
-      phoneNumber
-      registrationDate
-      role
-    }
-  }
-`;
-
-export const queryMeContext = gql`
-  query meContext {
-    item: meContext {
-      id
-      nickName
-      role
-      picture {
-        id
-        filename
-      }
-    }
-  }
-`;
-
-export const mutationSignOut = gql`
-  mutation userSignOut {
-    userSignOut
-  }
-`;
