+++ frontend/src/graphql/auth/mutationResetPassword.ts
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client";
+
+export const mutationResetPassword = gql`
+  mutation resetPassword($email: String!) {
+    resetPassword(email: $email)
+  }
+`;
