+++ frontend/src/graphql/auth/mutationSetPassword.ts
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client";
+
+export const mutationSetPassword = gql`
+  mutation setPassword($password: String!, $token: String!) {
+    setPassword(password: $password, token: $token)
+  }
+`;
