+++ frontend/src/graphql/auth/mutationSignOut.ts
@@ -1,0 +1,7 @@
+import { gql } from "@apollo/client";
+
+export const mutationSignOut = gql`
+  mutation userSignOut {
+    userSignOut
+  }
+`;
