+++ frontend/src/graphql/auth/mutationUserLogin.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const mutationUserLogin = gql`
+  mutation userLogin($data: UserLoginInput!) {
+    item: userLogin(data: $data) {
+      id
+      firstName
+    }
+  }
+`;
