+++ frontend/src/graphql/auth/mutationVerifyEmail.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const mutationVerifyEmail = gql`
+  mutation verifyEmail($token: String!) {
+    item: verifyEmail(token: $token) {
+      success
+      message
+    }
+  }
+`;
