+++ frontend/src/graphql/categories/mutationCreateCategory.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const mutationCreateCategory = gql`
+  mutation categoryCreate($data: CategoryCreateInput!) {
+    item: categoryCreate(data: $data) {
+      id
+    }
+  }
+`;
