+++ frontend/src/graphql/categories/queryAllCatWithHierarchy.ts
@@ -1,0 +1,18 @@
+import { gql } from "@apollo/client";
+
+export const queryAllCatWithHierarchy = gql`
+  query CategoriesGetAllWithHierarchy {
+    items: categoriesGetAllWithHierarchy {
+      id
+      name
+      childCategories {
+        id
+        name
+        childCategories {
+          id
+          name
+        }
+      }
+    }
+  }
+`;
