+++ frontend/src/graphql/categories/queryAllCategories.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const queryAllCategories = gql`
+  query CategoriesGetAll {
+    items: categoriesGetAll {
+      id
+      name
+    }
+  }
+`;
