+++ frontend/src/graphql/categories/queryAllRootCategories.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const queryAllRootCategories = gql`
+  query CategoriesGetaLLRoot {
+    items: categoriesGetaLLRoot {
+      id
+      name
+    }
+  }
+`;
