+++ frontend/src/graphql/users/mutationCreateUser.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const mutationCreateUser = gql`
+  mutation userCreate($data: UserCreateInput!) {
+    item: userCreate(data: $data) {
+      id
+      nickName
+    }
+  }
+`;
