+++ frontend/src/graphql/users/queryMeContext.ts
@@ -1,0 +1,15 @@
+import { gql } from "@apollo/client";
+
+export const queryMeContext = gql`
+  query meContext {
+    item: meContext {
+      id
+      nickName
+      role
+      picture {
+        id
+        filename
+      }
+    }
+  }
+`;
